refactor:Refactor files and project folder
