+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import '../src/assets/css/index.css';
 import App from './App';
 
 ReactDOM.render(
